fix(scripts): use separator constant, move computeBuildSizes into scripts/
